+++ CI/SCRIPTS/test_stac.py
@@ -58,7 +58,7 @@ from eoreader.utils import EOREADER_NAME
 
 from .scripts_utils import CI_EOREADER_S3, READER, dask_env, opt_path, s3_env, sar_path
 
-ci.reduce_verbosity()
+ci.reduce_verbosity(["dicttoxml"])
 logging.getLogger("rasterio._env").setLevel(logging.ERROR)
 
 LOGGER = logging.getLogger(EOREADER_NAME)
@@ -658,3 +658,10 @@ def test_iceye():
 def test_saocom():
     """Function testing the support of SAOCOM constellation"""
     _test_core_sar("*SAO*")
+
+
+@s3_env
+@dask_env
+def test_capella():
+    """Function testing the support of CAPELLA constellation"""
+    _test_core_sar("*CAPELLA*")